requirements.txt
psycopg2==2.9.9
Django==4.2.6
djangorestframework==3.14.0
django-cpf==0.1.0